+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+# 2.41.3
+
 * Performance optimizations to tile reading, writing, and overzooming
+* Automatically vary the tile size limit by zoom level to match the intended retain-points-multiplier multiplication
+* Fix decompression error when bailing out because a tile can't be made small enough
+* Search harder for a feature size threshold to make the tile small enough before giving up
 
 # 2.41.2